// 12042021/smart_front-end-main/src/app/models/artiste.ts
export class artiste {
  id: Number;
  cdate: Date;
   nom: String;
   prenom: String;
   cin: Number;
   contrat: String;
   date: Date;
   datecin: Date;
   email: String;
   image: String;
   nArtistique: String;
   nationnalite: String;
  part: Number;
   phone: Number;
  proposition: String;
  retenu: Number;

}